# swagger-pt.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WhatsApp API",
    "description": "API Wrapper para WhatsAppWebJS - Interface REST para interagir com o WhatsApp Web",
    "version": "1.0.0"
  },
  "host": "",
...
  "paths": {
...
    "/cache/status": {
      "get": {
        "summary": "Status do Cache",
        "description": "Obter status do sistema de cache e estatísticas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cache status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "cache": {
                      "type": "object",
                      "properties": {
                        "useRedis": {
                          "type": "boolean"
                        },
                        "redisConnected": {
                          "type": "boolean"
                        },
                        "memoryCacheSize": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/cache/clear": {
      "post": {
        "summary": "Limpar Cache",
        "description": "Limpar todos os dados do cache",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cache cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/localCallbackExample": {
      "post": {
        "summary": "Exemplo de Webhook Local",
        "description": "Example endpoint for testing webhooks locally",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallbackResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/session/start/{sessionId}": {
      "get": {
        "tags": [
          "Sessão"
        ],
        "summary": "Iniciar Nova Sessão",
        "description": "Inicia uma nova sessão para o ID de sessão fornecido.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier for the session (alphanumeric and - allowed)",
            "example": "f8377d8d-a589-4242-9ba6-9486a04ef80c"
          }
        ],
        "responses": {
          "200": {
            "description": "Status of the initiated session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartSessionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/session/status/{sessionId}": {
      "get": {
        "tags": [
          "Sessão"
        ],
        "summary": "Obter Status da Sessão",
        "description": "Status da sessão com o ID de sessão fornecido.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier for the session (alphanumeric and - allowed)",
            "example": "f8377d8d-a589-4242-9ba6-9486a04ef80c"
          }
        ],
        "responses": {
          "200": {
            "description": "Status of the session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusSessionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/session/qr/{sessionId}": {
      "get": {
        "tags": [
          "Sessão"
        ],
        "summary": "Obter Código QR da Sessão",
        "description": "Código QR da sessão com o ID de sessão fornecido. Retorna o código QR como uma string de texto que pode ser usada para gerar imagens QR ou para fins de depuração.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier for the session (alphanumeric and - allowed)",
            "example": "f8377d8d-a589-4242-9ba6-9486a04ef80c"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do código QR como string de texto",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "qr": {
                      "type": "string",
                      "description": "String de dados do código QR",
                      "example": "2@..."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/session/qr/{sessionId}/image": {
      "get": {
        "tags": [
          "Sessão"
        ],
        "summary": "Obter Código QR da Sessão como Imagem",
        "description": "Código QR como imagem da sessão com o ID de sessão fornecido. Retorna uma imagem PNG que pode ser exibida em navegadores ou aplicações. O código QR é gerado quando a sessão do WhatsApp requer autenticação.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier for the session (alphanumeric and - allowed)",
            "example": "f8377d8d-a589-4242-9ba6-9486a04ef80c"
          }
        ],
        "responses": {
          "200": {
            "description": "Imagem QR em formato PNG. A imagem pode ser exibida diretamente em navegadores ou salva em arquivo.",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                },
                "example": "iVBORw0KGgoAAAANSUhEUgAA..."
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/session/terminate/{sessionId}": {
      "get": {
        "tags": [
          "Sessão"
        ],
        "summary": "Encerrar Sessão",
        "description": "Encerra a sessão com o ID de sessão fornecido.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier for the session (alphanumeric and - allowed)",
            "example": "f8377d8d-a589-4242-9ba6-9486a04ef80c"
          }
        ],
        "responses": {
          "200": {
            "description": "Sessions terminated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminateSessionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/session/terminateInactive": {
      "get": {
        "tags": [
          "Sessão"
        ],
        "summary": "Encerrar Sessões Inativas",
        "description": "Encerra todas as sessões inativas.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sessions terminated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminateSessionsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/session/terminateAll": {
      "get": {
        "tags": [
          "Sessão"
        ],
        "summary": "Encerrar Todas as Sessões",
        "description": "Encerra todas as sessões.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sessions terminated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminateSessionsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "StartSessionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Session initiated successfully"
          }
        },
        "xml": {
          "name": "StartSessionResponse"
        }
      },
      "StatusSessionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "state": {
            "type": "string",
            "example": "CONNECTED"
          },
          "message": {
            "type": "string",
            "example": "session_connected"
          }
        },
        "xml": {
          "name": "StatusSessionResponse"
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Some server error"
          }
        },
        "xml": {
          "name": "ErrorResponse"
        }
      },
...
      "ForbiddenResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Invalid API key"
          }
        },
        "xml": {
          "name": "ForbiddenResponse"
        }
      }
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}